all gradient examples have been moved into one unified
