--- examples/gradient_radial.rs
@@ -1,19 +1,0 @@
-#![cfg_attr(
-    all(not(debug_assertions), target_os = "windows"),
-    windows_subsystem = "windows"
-)]
-
-use freya::prelude::*;
-
-fn main() {
-    launch(app);
-}
-
-fn app() -> Element {
-    rsx!(rect {
-        height: "100%",
-        width: "100%",
-        background:
-            "radial-gradient(orange 15%, rgb(255, 0, 0) 50%, rgb(255, 192, 203) 80%)",
-    })
-}
